+++ src/index.ts
@@ -118,6 +118,7 @@ export type {
   ValidateOnConfig,
   UseFormReturnType,
   UseFormConfiguration,
+  FormErrorInput,
   ValidationError,
   ValidationResponse,
   ValidationResponseWithoutValue,